requirement.txt
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn
pillow
tqdm